Separate toctree manager to sphinx.environment.managers.toctree
